refactored aws settings into separate fileg
