+++ api/src/settings/settings.py
@@ -1,3 +1,4 @@
 from settings.base import *
 from settings.cors import *
 from settings.logging import *
+from settings.aws import *
